% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_quote.R
\name{bb_quote}
\alias{bb_quote}
\title{Retrieve a Random Breaking Bad Quote}
\usage{
bb_quote(quote_id = NULL, quote_by = NULL)
}
\arguments{
\item{quote_id}{Look up a quote by id.}

\item{quote_by}{Look up a quote by character in the show.}
}
\value{
an S3 object
}
\description{
This function retrieves a random quote from
\url{https://breakingbadapi.com} and returns an S3 object containing the
parsed request result and the request itself.
}
