% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_quote.R
\name{one_who_knocks}
\alias{one_who_knocks}
\title{Print a Breaking Bad Quote and Author}
\usage{
one_who_knocks(quote_id = NULL, quote_by = NULL, message = TRUE)
}
\arguments{
\item{quote_id}{Look up a quote by id.}

\item{quote_by}{Look up a quote by character in the show.}

\item{message}{A boolean. If TRUE, the quote is messaged to console.}
}
\value{
a character vector
}
\description{
This function returns a random quote and its author from
\url{https://breakingbadapi.com} as a single character vector. It first
messages the quote and invisibly returns it.
}
