% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_quote.R
\name{bb_quote}
\alias{bb_quote}
\title{Retrieve a Random Breaking Bad Quote}
\usage{
bb_quote(quote_id = NULL)
}
\arguments{
\item{quote_id}{A specific quote ID to return}
}
\value{
an S3 object
}
\description{
This function retrieves a random quote from
\url{https://breakingbadapi.com} and returns an S3 object containing the
parsed request result and the request itself.
}
