% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_quote.R
\name{bb_quote}
\alias{bb_quote}
\title{Function to return a random quote}
\usage{
bb_quote()
}
\value{
Invisibly returns the first item holding the quote as a character.
}
\description{
This function returns a random quote from
\url{https://breakingbadapi.com}. It first messages the quote and invisibly
returns it as a character.
}
