% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_quote.R
\name{one_who_knocks}
\alias{one_who_knocks}
\title{Print a Breaking Bad Quote and Author}
\usage{
one_who_knocks()
}
\value{
a character vector
}
\description{
This function returns a random quote and its author from
\url{https://breakingbadapi.com} as a single character vector. It first
messages the quote and invisibly returns it.
}
